fix: DIA-1983: Decouple LLM compute / message parser / LiteLLM interface
