+++ adala/utils/parse.py
@@ -2,7 +2,8 @@ import re
 import string
 import logging
 from string import Formatter
-from typing import List, TypedDict
+from typing import List, TypedDict, Dict, Optional, Union, Literal, Iterable, Generator
+from enum import Enum
 
 logger = logging.getLogger(__name__)
 
@@ -168,3 +169,127 @@ def parse_template(string, include_texts=True) -> List[TemplateChunks]:
         )
 
     return chunks
+
+
+# TODO: consolidate these data models and unify our preprocessing for LLM input into one step RawInputModel -> PreparedInputModel
+class TextMessageChunk(TypedDict):
+    type: Literal["text"]
+    text: str
+
+
+class ImageMessageChunk(TypedDict):
+    type: Literal["image"]
+    image_url: Dict[str, str]
+
+
+MessageChunk = Union[TextMessageChunk, ImageMessageChunk]
+
+Message = Union[str, List[MessageChunk]]
+
+
+class MessageChunkType(Enum):
+    TEXT = "text"
+    IMAGE_URL = "image_url"
+
+
+def split_message_into_chunks(
+    input_template: str, input_field_types: Dict[str, MessageChunkType], **input_fields
+) -> List[MessageChunk]:
+    """Split a template string with field types into a list of message chunks.
+
+    Takes a template string with placeholders and splits it into chunks based on the field types,
+    preserving the text between placeholders.
+
+    Args:
+        input_template (str): Template string with placeholders, e.g. '{a} is a {b} is an {a}'
+        input_field_types (Dict[str, MessageChunkType]): Dict mapping field names to their types
+        **input_fields: Field values to substitute into template
+
+    Returns:
+        List[Dict[str, str]]: List of message chunks with appropriate type and content.
+            Text chunks have format: {'type': 'text', 'text': str}
+            Image chunks have format: {'type': 'image_url', 'image_url': {'url': str}}
+
+    Example:
+        >>> split_message_into_chunks(
+        ...     '{a} is a {b} is an {a}',
+        ...     {'a': MessageChunkType.TEXT, 'b': MessageChunkType.IMAGE_URL},
+        ...     a='the letter a',
+        ...     b='http://example.com/b.jpg'
+        ... )
+        [
+            {'type': 'text', 'text': 'the letter a is a '},
+            {'type': 'image_url', 'image_url': {'url': 'http://example.com/b.jpg'}},
+            {'type': 'text', 'text': ' is an the letter a'}
+        ]
+    """
+    # Parse template to get field positions and surrounding text
+    parsed = parse_template(input_template)
+
+    def add_to_current_chunk(
+        current_chunk: Optional[MessageChunk], chunk: MessageChunk
+    ) -> MessageChunk:
+        if current_chunk:
+            current_chunk["text"] += chunk["text"]
+            return current_chunk
+        else:
+            return chunk
+
+    # Build chunks by iterating through parsed template parts
+    def build_chunks(
+        parsed: Iterable[TemplateChunks],
+    ) -> Generator[MessageChunk, None, None]:
+        current_chunk: Optional[MessageChunk] = None
+
+        for part in parsed:
+            if part["type"] == "text":
+                current_chunk = add_to_current_chunk(
+                    current_chunk, {"type": "text", "text": part["text"]}
+                )
+            elif part["type"] == "var":
+                field_value = part["text"]
+                try:
+                    field_type = input_field_types[field_value]
+                except KeyError:
+                    raise ValueError(
+                        f"Field {field_value} not found in input_field_types. Found fields: {input_field_types}"
+                    )
+                if field_type == MessageChunkType.TEXT:
+                    # try to substitute in variable and add to current chunk
+                    substituted_text = partial_str_format(
+                        f"{{{field_value}}}", **input_fields
+                    )
+                    if substituted_text != field_value:
+                        current_chunk = add_to_current_chunk(
+                            current_chunk, {"type": "text", "text": substituted_text}
+                        )
+                    else:
+                        # be permissive for unfound variables
+                        current_chunk = add_to_current_chunk(
+                            current_chunk,
+                            {"type": "text", "text": f"{{{field_value}}}"},
+                        )
+                elif field_type == MessageChunkType.IMAGE_URL:
+                    substituted_text = partial_str_format(
+                        f"{{{field_value}}}", **input_fields
+                    )
+                    if substituted_text != field_value:
+                        # push current chunk, push image chunk, and start new chunk
+                        if current_chunk:
+                            yield current_chunk
+                        current_chunk = None
+                        yield {
+                            "type": "image_url",
+                            "image_url": {"url": input_fields[field_value]},
+                        }
+                    else:
+                        # be permissive for unfound variables
+                        current_chunk = add_to_current_chunk(
+                            current_chunk,
+                            {"type": "text", "text": f"{{{field_value}}}"},
+                        )
+
+        if current_chunk:
+            yield current_chunk
+
+    return list(build_chunks(parsed))
